introduce a new module "note.rs" and move backend functionality there
